+++ plugins/node/opentelemetry-instrumentation-aws-lambda/src/instrumentation.ts
@@ -53,13 +53,10 @@ import {
   Handler,
 } from 'aws-lambda';
 
-import {
-  LambdaModule,
-  AwsLambdaInstrumentationConfig,
-  EventContextExtractor,
-} from './types';
+import { AwsLambdaInstrumentationConfig, EventContextExtractor } from './types';
 import { VERSION } from './version';
 import { env } from 'process';
+import { LambdaModule } from './internal-types';
 
 const awsPropagator = new AWSXRayPropagator();
 const headerGetter: TextMapGetter<APIGatewayProxyEventHeaders> = {