+++ plugins/node/opentelemetry-instrumentation-connect/src/index.ts
@@ -16,4 +16,3 @@
 
 export * from './enums/AttributeNames';
 export * from './instrumentation';
-export * from './types';