+++ plugins/node/opentelemetry-instrumentation-graphql/src/types.ts
@@ -15,21 +15,10 @@
  */
 
 import { InstrumentationConfig } from '@opentelemetry/instrumentation';
-import type * as graphqlTypes from 'graphql';
 import type * as api from '@opentelemetry/api';
-import type { PromiseOrValue } from 'graphql/jsutils/PromiseOrValue';
-import { DocumentNode } from 'graphql/language/ast';
-import {
-  GraphQLFieldResolver,
-  GraphQLTypeResolver,
-} from 'graphql/type/definition';
-import { OTEL_GRAPHQL_DATA_SYMBOL, OTEL_PATCHED_SYMBOL } from './symbols';
-
-export const OPERATION_NOT_SUPPORTED =
-  'Operation$operationName$not' + ' supported';
 
 export interface GraphQLInstrumentationExecutionResponseHook {
-  (span: api.Span, data: graphqlTypes.ExecutionResult): void;
+  (span: api.Span, data: any): void;
 }
 
 export interface GraphQLInstrumentationConfig extends InstrumentationConfig {
@@ -41,6 +30,7 @@ export interface GraphQLInstrumentationConfig extends InstrumentationConfig {
    * @default false
    */
   allowValues?: boolean;
+
   /**
    * The maximum depth of fields/resolvers to instrument.
    * When set to 0 it will not instrument fields and resolvers
@@ -48,6 +38,19 @@ export interface GraphQLInstrumentationConfig extends InstrumentationConfig {
    * @default undefined
    */
   depth?: number;
+
+  /**
+   * Don't create spans for the execution of the default resolver on object properties.
+   *
+   * When a resolver function is not defined on the schema for a field, graphql will
+   * use the default resolver which just looks for a property with that name on the object.
+   * If the property is not a function, it's not very interesting to trace.
+   * This option can reduce noise and number of spans created.
+   *
+   * @default false
+   */
+  ignoreTrivialResolveSpans?: boolean;
+
   /**
    * Whether to merge list items into a single element.
    *
@@ -61,100 +64,12 @@ export interface GraphQLInstrumentationConfig extends InstrumentationConfig {
    * Hook that allows adding custom span attributes based on the data
    * returned from "execute" GraphQL action.
    *
+   * @param data - A GraphQL `ExecutionResult` object. For the exact type definitions, see the following:
+   *  - {@linkcode https://github.com/graphql/graphql-js/blob/v14.7.0/src/execution/execute.js#L115 graphql@14}
+   *  - {@linkcode https://github.com/graphql/graphql-js/blob/15.x.x/src/execution/execute.d.ts#L31 graphql@15}
+   *  - {@linkcode https://github.com/graphql/graphql-js/blob/16.x.x/src/execution/execute.ts#L127 graphql@16}
+   *
    * @default undefined
    */
   responseHook?: GraphQLInstrumentationExecutionResponseHook;
 }
-
-/**
- * Merged and parsed config of default instrumentation config and GraphQL
- */
-export type GraphQLInstrumentationParsedConfig =
-  Required<GraphQLInstrumentationConfig> & InstrumentationConfig;
-
-export type executeFunctionWithObj = (
-  args: graphqlTypes.ExecutionArgs
-) => PromiseOrValue<graphqlTypes.ExecutionResult>;
-
-export type executeArgumentsArray = [
-  graphqlTypes.GraphQLSchema,
-  graphqlTypes.DocumentNode,
-  any,
-  any,
-  Maybe<{ [key: string]: any }>,
-  Maybe<string>,
-  Maybe<graphqlTypes.GraphQLFieldResolver<any, any>>,
-  Maybe<graphqlTypes.GraphQLTypeResolver<any, any>>
-];
-
-export type executeFunctionWithArgs = (
-  schema: graphqlTypes.GraphQLSchema,
-  document: graphqlTypes.DocumentNode,
-  rootValue?: any,
-  contextValue?: any,
-  variableValues?: Maybe<{ [key: string]: any }>,
-  operationName?: Maybe<string>,
-  fieldResolver?: Maybe<graphqlTypes.GraphQLFieldResolver<any, any>>,
-  typeResolver?: Maybe<graphqlTypes.GraphQLTypeResolver<any, any>>
-) => PromiseOrValue<graphqlTypes.ExecutionResult>;
-
-export interface OtelExecutionArgs {
-  schema: graphqlTypes.GraphQLSchema;
-  document: DocumentNode & ObjectWithGraphQLData;
-  rootValue?: any;
-  contextValue?: any & ObjectWithGraphQLData;
-  variableValues?: Maybe<{ [key: string]: any }>;
-  operationName?: Maybe<string>;
-  fieldResolver?: Maybe<GraphQLFieldResolver<any, any> & OtelPatched>;
-  typeResolver?: Maybe<GraphQLTypeResolver<any, any>>;
-}
-
-export type executeType = executeFunctionWithObj | executeFunctionWithArgs;
-
-export type parseType = (
-  source: string | graphqlTypes.Source,
-  options?: graphqlTypes.ParseOptions
-) => graphqlTypes.DocumentNode;
-
-export type validateType = (
-  schema: graphqlTypes.GraphQLSchema,
-  documentAST: graphqlTypes.DocumentNode,
-  rules?: ReadonlyArray<graphqlTypes.ValidationRule>,
-  options?: { maxErrors?: number },
-  typeInfo?: graphqlTypes.TypeInfo
-) => ReadonlyArray<graphqlTypes.GraphQLError>;
-
-export interface GraphQLField {
-  parent: api.Span;
-  span: api.Span;
-  error: Error | null;
-}
-
-interface OtelGraphQLData {
-  source?: any;
-  span: api.Span;
-  fields: { [key: string]: GraphQLField };
-}
-
-export interface ObjectWithGraphQLData {
-  [OTEL_GRAPHQL_DATA_SYMBOL]?: OtelGraphQLData;
-}
-
-export interface OtelPatched {
-  [OTEL_PATCHED_SYMBOL]?: boolean;
-}
-
-export interface GraphQLPath {
-  prev: GraphQLPath | undefined;
-  key: string | number;
-  /**
-   * optional as it didn't exist yet in ver 14
-   */
-  typename?: string | undefined;
-}
-
-/**
- * Moving this type from ver 15 of graphql as it is nto available in ver. 14s
- * this way it can compile against ver 14.
- */
-export type Maybe<T> = null | undefined | T;