+++ plugins/node/opentelemetry-instrumentation-graphql/CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+## [0.33.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-graphql-v0.32.0...instrumentation-graphql-v0.33.0) (2022-11-16)
+
+
+### Features
+
+* **graphql:** add ignoreTrivialResolveSpans config option ([#1256](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/1256)) ([aff84bb](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/aff84bba1d391ec2061b8d0121ac8dd36fc1980c))
+* update experimental deps to `^0.34.0`, core deps to `^1.8.0`, api to `^1.3.0` ([#1278](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/1278)) ([e9fe8e1](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/e9fe8e13e34f54e96c50525cadeb74ac048c5624))
+
+
+### Bug Fixes
+
+* **instrumentation-graphql:** stop using types from `graphql` library in public api ([#1268](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/1268)) ([f8cabf3](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/f8cabf306faa3e3eb4b9ce38ccdde842abdb2b82))
+
+## [0.32.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-graphql-v0.31.0...instrumentation-graphql-v0.32.0) (2022-11-02)
+
+
+### Features
+
+* support `graphql` v16 ([#998](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/998)) ([5da46ef](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/5da46ef7a29bbc64f600d794b1e68bb6738a9f2e))
+
+
+### Bug Fixes
+
+* **graphql:** graphql instrumentation throw for sync calls ([#1254](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/1254)) ([524d98e](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/524d98e4fac3322b7da3cc865f53043f03f67bb7))
+* separate public and internal types for all instrumentations ([#1251](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/1251)) ([e72ea58](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/e72ea58cfb888a90590970f63d3a042a8ea3aaf2))
+
 ## [0.31.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-graphql-v0.30.0...instrumentation-graphql-v0.31.0) (2022-09-15)